// lib/api/generated/model/unitDetailDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API Documentation
 * The API description
 * OpenAPI spec version: 1.0
 */
import type { UnitDetailDtoLevel } from './unitDetailDtoLevel';
import type { UnitDetailDtoEpisodesItem } from './unitDetailDtoEpisodesItem';

export interface UnitDetailDto {
  id: number;
  title: string;
  description: string;
  level: UnitDetailDtoLevel;
  episodes: UnitDetailDtoEpisodesItem[];
}