// lib/api/generated/model/unitListItemDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API Documentation
 * The API description
 * OpenAPI spec version: 1.0
 */
import type { UnitListItemDtoStory } from './unitListItemDtoStory';

export interface UnitListItemDto {
  id: number;
  title: string;
  description: string;
  color: string;
  order: number;
  story: UnitListItemDtoStory;
}